backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
pydantic-settings==2.1.0
python-dotenv>=0.19.0
email-validator>=1.1.3
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
httpx>=0.24.1
sqlalchemy==2.0.23
asyncpg==0.29.0
pymongo==4.5.0
motor==3.3.0
redis==5.0.1
aioredis==2.0.1
aio-pika==9.3.0
pytest==7.4.3
pytest-asyncio==0.21.1
# External APIs
asana==3.2.1
twilio==8.10.1
# Email management
mailjet-rest==1.3.4

# Task queue and celery
celery==5.3.4
flower==2.0.1

# Monitoring and logging
prometheus-client==0.18.0
python-json-logger==2.0.7

# Web scraping and data analysis
beautifulsoup4==4.12.2
instaloader==4.10.0

# Social media management
linkedin-api==2.0.0a5
praw==7.7.1
facebook-business==17.0.2
instabot==0.117.0
tiktok-uploader==1.0.4
moviepy==1.0.3
Pillow==10.1.0
opencv-python==4.8.1.78
pandas==2.1.3

# Additional utilities
python-dateutil==2.8.2

# AI and machine learning
google-api-python-client>=2.86.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
minio>=7.1.15
requests>=2.31.0

# Testing
